requirements.txt
langchain-core
langchain-community
python-dotenv
streamlit
langserve
fastapi